# api/api-docs.yaml
openapi: 3.0.3
info:
  title: MicroMentor API
  version: 1.0.0
  description: |
    REST API for MicroMentor — a knowledge marketplace connecting users with verified
    expert professionals for short, high-impact online consultation sessions.

    ## Authentication
    All protected endpoints require a Firebase ID token passed as a Bearer token:
    ```
    Authorization: Bearer <firebase-id-token>
    ```
    Firebase handles registration, login, Google OAuth, and password reset.
    The backend verifies the token and manages local user records.

    ## Roles
    - `USER` — requests consultations
    - `EXPERT` — provides consultation services (a user with an approved expert application)
    - `ADMIN` — platform administrator with a separate admin profile

    ## Currency
    All monetary values are in Vietnamese Dong (VND), stored and transmitted as integers.
servers:
  - url: http://localhost:8000/api
    description: Local development
  - url: https://api.micromentor.vn/api
    description: Production
...
components:
  securitySchemes:
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: Firebase ID token obtained from the Firebase client SDK.

  # ─────────────────────────────────────────────
  # REUSABLE PARAMETERS
  # ─────────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    ExpertIdPath:
      name: expertId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ExpertCategory:
      type: string
      enum:
        - technology
        - business
        - finance
        - marketing
        - design
        - legal
        - healthcare
        - education
        - career
        - startup
    SessionDuration:
      type: integer
      enum: [30, 60, 90]
      description: Session duration in minutes (maps to 2, 4, or 6 consecutive 15-min slots)
...
    ExpertProfile:
      type: object
      required:
        - id
        - userId
        - slug
        - displayName
        - title
        - bio
        - skills
        - yearsOfExperience
        - pricePerSession
        - rating
        - reviewCount
        - totalSessions
        - category
        - isAvailable
        - profileStatus
      properties:
        id:
          type: string
          example: exp_01HXYZ
        userId:
          type: string
          example: usr_01HXYZ
        slug:
          type: string
          example: nguyen-van-a-tech
        displayName:
          type: string
          example: Nguyễn Văn A
        title:
          type: string
          example: Senior Software Engineer
        company:
          type: string
          nullable: true
          example: Google Vietnam
        bio:
          type: string
          example: >
            Kỹ sư phần mềm với hơn 10 năm kinh nghiệm trong lĩnh vực công nghệ.
        skills:
          type: array
          items:
            type: string
          example: [React, Node.js, Python, AWS]
        yearsOfExperience:
          type: integer
          example: 10
        pricePerSession:
          type: integer
          description: Price per 15-minute slot in VND
          example: 300000
        sessionDurationMinutes:
          type: integer
          description: Preferred session length in minutes (30, 60, or 90)
          example: 60
        rating:
          type: number
          format: float
          example: 4.8
        reviewCount:
          type: integer
          example: 42
        totalSessions:
          type: integer
          example: 105
        linkedinUrl:
          type: string
          format: uri
          nullable: true
          example: https://linkedin.com/in/nguyenvana
        portfolioUrl:
          type: string
          format: uri
          nullable: true
        certifications:
          type: array
          items:
            $ref: '#/components/schemas/Certification'
        portfolio:
          type: array
          items:
            $ref: '#/components/schemas/PortfolioItem'
        languages:
          type: array
          items:
            type: string
          example: [vi, en]
        category:
          $ref: '#/components/schemas/ExpertCategory'
        isAvailable:
          type: boolean
          example: true
        profilePictureUrl:
          type: string
          format: uri
          nullable: true
        profileStatus:
          $ref: '#/components/schemas/ExpertProfileStatus'
        totalEarnings:
          type: integer
          description: Lifetime earnings in VND (only visible to the expert themselves)
          example: 15000000
        pendingBalance:
          type: integer
          description: Unpaid balance in VND (only visible to the expert themselves)
          example: 2400000
...
    ExpertListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExpertProfile'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
  responses:
...
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: Resource not found.
...
paths:
...
  /experts:
    get:
      summary: Search and list experts
      description: |
        Publicly accessible. Returns approved experts matching the given filters.
        Supports full-text search across name, title, bio, and skills.
      operationId: listExperts
      tags: [Experts]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: q
          in: query
          description: Full-text search query
          schema:
            type: string
            example: react
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/ExpertCategory'
        - name: minPrice
          in: query
          schema:
            type: integer
            example: 0
        - name: maxPrice
          in: query
          schema:
            type: integer
            example: 1000000
        - name: minRating
          in: query
          schema:
            type: number
            format: float
            example: 4.0
        - name: languages
          in: query
          schema:
            type: string
            description: Comma-separated language codes
            example: vi,en
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [rating_desc, price_asc, price_desc, experience_desc, sessions_desc]
            default: rating_desc
      responses:
        '200':
          description: Paginated list of experts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertListResponse'
  /experts/{expertId}:
    get:
      summary: Get expert public profile
      description: Returns the full public profile of an approved expert by ID or slug.
      operationId: getExpert
      tags: [Experts]
      parameters:
        - $ref: '#/components/parameters/ExpertIdPath'
      responses:
        '200':
          description: Expert profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertProfile'
        '404':
          $ref: '#/components/responses/NotFound'
  /experts/{expertId}/reviews:
    get:
      summary: Get reviews for an expert
      description: Returns paginated public reviews for the given expert.
      operationId: listExpertReviews
      tags: [Experts]
      parameters:
        - $ref: '#/components/parameters/ExpertIdPath'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated reviews.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /experts/{expertId}/availability:
    get:
      summary: Get expert public availability
      description: Returns available (unbooked) slots for the given expert within a date range.
      operationId: getExpertPublicAvailability
      tags: [Experts]
      parameters:
        - $ref: '#/components/parameters/ExpertIdPath'
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2025-06-01"
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2025-06-30"
      responses:
        '200':
          description: Available slots grouped by date.
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      $ref: '#/components/schemas/AvailabilitySlot'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────────
  # EXPERT APPLICATIONS
  # ─────────────────────────────────────────────
  /expert-applications:
    post:
      summary: Submit expert application
      description: |
        Allows an authenticated USER to apply to become an expert.
        A user may only have one active application at a time.
      operationId: submitExpertApplication
      tags: [Expert Applications]
      security:
        - FirebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - yearsOfExperience
                - category
                - bio
                - pricePerSession
                - skills
              properties:
                title:
                  type: string
                  minLength: 3
                  example: Senior Software Engineer
                company:
                  type: string
                  nullable: true
                  example: Google Vietnam
                yearsOfExperience:
                  type: integer
                  minimum: 1
                  maximum: 50
                  example: 8
                category:
                  $ref: '#/components/schemas/ExpertCategory'
                bio:
                  type: string
                  minLength: 100
                  example: "Tôi là kỹ sư phần mềm với hơn 8 năm kinh nghiệm..."
                pricePerSession:
                  type: integer
                  minimum: 100000
                  example: 200000
                skills:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  example: [React, Node.js, AWS]
                linkedinUrl:
                  type: string
                  format: uri
                  nullable: true
                portfolioUrl:
                  type: string
                  format: uri
                  nullable: true
                certifications:
                  type: array
                  items:
                    $ref: '#/components/schemas/Certification'
      responses:
        '201':
          description: Application submitted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertApplication'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /expert-applications/me:
    get:
      summary: Get current user's expert application
      description: Returns the authenticated user's most recent expert application.
      operationId: getMyExpertApplication
      tags: [Expert Applications]
      security:
        - FirebaseAuth: []
      responses:
        '200':
          description: Expert application.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertApplication'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update expert application
      description: |
        Update a `NEEDS_REVISION` application with revised information.
        Only allowed when status is `NEEDS_REVISION`.
      operationId: updateMyExpertApplication
      tags: [Expert Applications]
      security:
        - FirebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                bio:
                  type: string
                  minLength: 100
                skills:
                  type: array
                  items:
                    type: string
                pricePerSession:
                  type: integer
                  minimum: 100000
                certifications:
                  type: array
                  items:
                    $ref: '#/components/schemas/Certification'
      responses:
        '200':
          description: Application updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertApplication'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      summary: Withdraw expert application
      description: Withdraw a `PENDING_REVIEW` application. Cannot be undone.
      operationId: withdrawMyExpertApplication
      tags: [Expert Applications]
      security:
        - FirebaseAuth: []
      responses:
        '204':
          description: Application withdrawn.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'

  # ─────────────────────────────────────────────
  # EXPERT PROFILE MANAGEMENT (own profile)
  # ─────────────────────────────────────────────
  /expert/profile:
    get:
      summary: Get current expert's own profile
      description: Returns the full expert profile including private fields (earnings, balance).
      operationId: getMyExpertProfile
      tags: [Expert Profile]
      security:
        - FirebaseAuth: []
      responses:
        '200':
          description: Expert profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    patch:
      summary: Update expert profile
      description: |
        Update the expert's own profile. Significant changes (bio, title, pricing)
        will set `profileStatus` to `PENDING_REVIEW` until an admin approves them.
      operationId: updateMyExpertProfile
      tags: [Expert Profile]
      security:
        - FirebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                title:
                  type: string
                company:
                  type: string
                  nullable: true
                bio:
                  type: string
                  minLength: 100
                skills:
                  type: array
                  items:
                    type: string
                pricePerSession:
                  type: integer
                  minimum: 100000
                sessionDurationMinutes:
                  $ref: '#/components/schemas/SessionDuration'
                languages:
                  type: array
                  items:
                    type: string
                isAvailable:
                  type: boolean
                linkedinUrl:
                  type: string
                  format: uri
                  nullable: true
      responses:
        '200':
          description: Updated expert profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /expert/profile/portfolio:
    post:
      summary: Add portfolio item
      description: Add a new portfolio item to the expert's profile.
      operationId: addPortfolioItem
      tags: [Expert Profile]
      security:
        - FirebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description]
              properties:
                title:
                  type: string
                description:
                  type: string
                url:
                  type: string
                  format: uri
                  nullable: true
                imageUrl:
                  type: string
                  format: uri
                  nullable: true
      responses:
        '201':
          description: Portfolio item added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /expert/profile/portfolio/{itemId}:
    patch:
      summary: Update portfolio item
      operationId: updatePortfolioItem
      tags: [Expert Profile]
      security:
        - FirebaseAuth: []
      parameters:
        - $ref: '#/components/parameters/PortfolioItemIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                url:
                  type: string
                  format: uri
                  nullable: true
      responses:
        '200':
          description: Portfolio item updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete portfolio item
      operationId: deletePortfolioItem
      tags: [Expert Profile]
      security:
        - FirebaseAuth: []
      parameters:
        - $ref: '#/components/parameters/PortfolioItemIdPath'
      responses:
        '204':
          description: Portfolio item deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /expert/profile/certifications:
    post:
      summary: Add certification
      operationId: addCertification
      tags: [Expert Profile]
      security:
        - FirebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, issuer, year]
              properties:
                name:
                  type: string
                issuer:
                  type: string
                year:
                  type: integer
                url:
                  type: string
                  format: uri
                  nullable: true
      responses:
        '201':
          description: Certification added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Certification'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /expert/profile/certifications/{certId}:
    delete:
      summary: Delete certification
      operationId: deleteCertification
      tags: [Expert Profile]
      security:
        - FirebaseAuth: []
      parameters:
        - $ref: '#/components/parameters/CertificationIdPath'
      responses:
        '204':
          description: Certification deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────────
  # AVAILABILITY
  # ─────────────────────────────────────────────
  /expert/availability:
    get:
      summary: Get own availability slots
      description: Returns all slots (booked and available) for the authenticated expert.
      operationId: getMyAvailability
      tags: [Availability]
      security:
        - FirebaseAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2025-06-01"
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2025-06-30"
      responses:
        '200':
          description: Availability slots.
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      $ref: '#/components/schemas/AvailabilitySlot'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create availability slots
      description: |
        Create one or more 15-minute availability slots. Each slot must have a
        unique `(expertId, date, startTime)` combination.
      operationId: createAvailabilitySlots
      tags: [Availability]
      security:
        - FirebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slots]
              properties:
                slots:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [date, startTime]
                    properties:
                      date:
                        type: string
                        format: date
                        example: "2025-06-01"
                      startTime:
                        type: string
                        pattern: '^([01]\d|2[0-3]):[0-5]\d$'
                        example: "09:00"
      responses:
        '201':
          description: Slots created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      $ref: '#/components/schemas/AvailabilitySlot'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /expert/availability/{slotId}:
    patch:
      summary: Update availability slot
      description: Update metadata on a slot. Cannot modify a slot that is already booked.
      operationId: updateAvailabilitySlot
      tags: [Availability]
      security:
        - FirebaseAuth: []
      parameters:
        - $ref: '#/components/parameters/SlotIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date
                startTime:
                  type: string
                  pattern: '^([01]\d|2[0-3]):[0-5]\d$'
      responses:
        '200':
          description: Slot updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilitySlot'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      summary: Delete availability slot
      description: Delete a slot. Cannot delete a booked slot.
      operationId: deleteAvailabilitySlot
      tags: [Availability]
      security:
        - FirebaseAuth: []
      parameters:
        - $ref: '#/components/parameters/SlotIdPath'
      responses:
        '204':
          description: Slot deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  # ─────────────────────────────────────────────
  # BOOKINGS
  # ─────────────────────────────────────────────
...
  /expert/payouts:
    get:
      summary: List expert payout history
      description: Returns paginated payout history for the authenticated expert.
      operationId: listMyPayouts
      tags: [Payouts]
      security:
        - FirebaseAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated payouts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayoutListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Request payout
      description: |
        Expert requests a payout for their pending balance. Creates a `PENDING`
        payout for admin processing.
      operationId: requestPayout
      tags: [Payouts]
      security:
        - FirebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, bankAccount]
              properties:
                amount:
                  type: integer
                  description: Amount in VND to withdraw (must not exceed pendingBalance)
                  example: 2400000
                bankAccount:
                  type: object
                  required: [bankName, accountNumber, accountName]
                  properties:
                    bankName:
                      type: string
                      example: Vietcombank
                    accountNumber:
                      type: string
                      example: "9876543210"
                    accountName:
                      type: string
                      example: NGUYEN VAN A
      responses:
        '201':
          description: Payout request created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payout'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /expert/payouts/summary:
    get:
      summary: Get expert earnings summary
      description: Returns total earnings, pending balance, and processed payout total.
      operationId: getPayoutSummary
      tags: [Payouts]
      security:
        - FirebaseAuth: []
      responses:
        '200':
          description: Earnings summary.
          content:
            application/json:
              schema:
                type: object
                required: [totalEarnings, pendingBalance, totalPaidOut]
                properties:
                  totalEarnings:
                    type: integer
                    example: 15000000
                  pendingBalance:
                    type: integer
                    example: 2400000
                  totalPaidOut:
                    type: integer
                    example: 12600000
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ═════════════════════════════════════════════
  # ADMIN PATHS
  # ═════════════════════════════════════════════
...